chore: skip the tests that use TTY on Windows
Our mock console library does not support PTY on Windows. So skip the tests for now.

Closes #59.
