+++ internal/pkg/cli/env_add_test.go
@@ -6,85 +6,12 @@ package cli
 import (
 	"testing"
 
-	"github.com/AlecAivazis/survey/v2/terminal"
-	"github.com/Netflix/go-expect"
 	"github.com/aws/PRIVATE-amazon-ecs-archer/internal/pkg/archer"
 	cli_mocks "github.com/aws/PRIVATE-amazon-ecs-archer/internal/pkg/cli/mocks"
 	"github.com/aws/PRIVATE-amazon-ecs-archer/mocks"
 	"github.com/golang/mock/gomock"
-	"github.com/hinshun/vt10x"
 	"github.com/stretchr/testify/require"
 )
-
-func TestEnvAdd_Ask(t *testing.T) {
-	testCases := map[string]struct {
-		inputEnv     string
-		inputProject string
-		input        func(c *expect.Console)
-
-		wantedEnv     string
-		wantedProject string
-	}{
-		"with no flags set": {
-			input: func(c *expect.Console) {
-				c.ExpectString("What is your project's name?")
-				c.SendLine("project")
-				c.ExpectString("What is your environment's name?")
-				c.SendLine("env")
-				c.ExpectEOF()
-			},
-
-			wantedEnv:     "env",
-			wantedProject: "project",
-		},
-		"with env flags set": {
-			input: func(c *expect.Console) {
-				c.ExpectString("What is your project's name?")
-				c.SendLine("project")
-				c.ExpectEOF()
-			},
-			inputEnv:      "env",
-			wantedEnv:     "env",
-			wantedProject: "project",
-		},
-	}
-
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			// GIVEN
-			mockTerminal, _, _ := vt10x.NewVT10XConsole()
-			defer mockTerminal.Close()
-			addEnv := &AddEnvOpts{
-				EnvName:     tc.inputEnv,
-				ProjectName: tc.inputProject,
-				prompt: terminal.Stdio{
-					In:  mockTerminal.Tty(),
-					Out: mockTerminal.Tty(),
-					Err: mockTerminal.Tty(),
-				},
-			}
-
-			// Write inputs to the terminal
-			done := make(chan struct{})
-			go func() {
-				defer close(done)
-				tc.input(mockTerminal)
-			}()
-
-			// WHEN
-			err := addEnv.Ask()
-
-			// Wait until the terminal receives the input
-			mockTerminal.Tty().Close()
-			<-done
-
-			// THEN
-			require.NoError(t, err)
-			require.Equal(t, tc.wantedProject, addEnv.ProjectName, "expected project names to match")
-			require.Equal(t, tc.wantedEnv, addEnv.EnvName, "expected environment names to match")
-		})
-	}
-}
 
 func TestEnvAdd_Validate(t *testing.T) {
 	testCases := map[string]struct {
@@ -116,7 +43,6 @@ func TestEnvAdd_Validate(t *testing.T) {
 		})
 	}
 }
-
 func TestEnvAdd_Execute(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	mockEnvStore := mocks.NewMockEnvironmentStore(ctrl)
